Accordion panels

Similar to collapse, accordion is a way to hide content. An accordion is made up of multiple panels and at most one panel is open. There is one exception, when the page renders multiple panels may start in the open state, however once the user interacts with accordion the default behavior returns. panel is a helper function to facilitate building accordion panels and does not have an explicit corresponding bootstrap class, see details for more information.

accordion(...)

accordionPanel(header, body = NULL, collapsed = TRUE, ...)

Arguments

...
Any number of accordionPanels or named arguments passed as HTML attributes to the parent element.

header
A character string specifying the header of an accordion panel.

body
A character string or tag element(s) specifying the body content of the panel, defaults to NULL.

collapsed
If TRUE, the accordion panel renders in the collapsed state, defaults to TRUE.

Details

accordion will generate an HTML id if an id argument is not passed in ... and child panels are assigned IDs based on the id of the accordion parent.

See also

For more information on accordion's, please refer to https://v4-alpha.getbootstrap.com/components/collapse/.

Examples

accordion(
  panel(
    heading = "Group 1",
    "Body text of the first group."
  ),
  panel(
    heading = "Group 2",
    "Body text of the second group."
  ),
  panel(
    heading = "Group 3",
    "Body text of the third group."
  )
)

#> Error in panel(heading = "Group 1", "Body text of the first group."): could not find function "panel"

if (interactive()) {
  shinyApp(
    ui = container(
      accordion(
        accordionPanel(
          "Option 1",
          "What do you think of option 1?",
          collapse = FALSE
        ),
        accordionPanel(
          "Option 2",
          "What do you think of option 2?"
        ),
        accordionPanel(
          "Option 3",
          "What do you think of option 3?"
        )
      )
    ),
    server = function(input, output) {

    }
  )
}