Buttons and button groups

Buttons, submit and reset buttons. A button's reactive value is a list of two items. The first item is count, the number of clicks on the button. The second item is value, the HTML data-value attribute of the button which may be set with the value argument.

button(label = NULL, value = NULL, context = "secondary",
  outline = FALSE, block = FALSE, disabled = FALSE, ..., id = NULL)

submit(label = NULL, outline = FALSE, block = FALSE, ...)

updateButton(id, count = NULL, context = NULL,
  session = getDefaultReactiveDomain())

Arguments

label
A character string or tag elements to use as a button label or a character vector specifying labels for a button group, defaults to NULL.

value
A character string specifying a value for the button or a character vector specifying values for a button group, defaults to NULL.

context
Used to specify the visual context of the button, one of "primary", "secondary", "success", "info", "warning", "danger", or "link", defaults to "secondary".

Primary buttons are blue, secondary buttons are white and grey, buttons for success are green, informative buttons are a lighter blue, warning buttons are yellow, and buttons for danger are red. Specifying "link" makes a button render with the appearance of a link.

outline
If TRUE, the button's background is transparent, context is preserved, defaults to FALSE.

block
If TRUE, the button is block-level instead of inline, defaults to FALSE. A block-level element will occupy the entire space of its parent element, thereby creating a "block."

disabled
If TRUE, the button renders in a disabled state, defaults to FALSE.

...
Named arguments passed as HTML attributes to the parent element.

id
A character string specifying the id of the button, button group, or submit button, defaults to NULL. If specified, a reactive value is available to the shiny server function.

See also

For more about block-level elements please refer to the block-level elements MDN reference section.

For more about buttons and button groups please refer to the bootstrap reference section.

Other inputs: groupInput, radioInput, selectInput, textInput

Examples

button("Primary", context = "primary")

#> <button class="dull-button dull-input btn btn-primary" type="button" role="button" data-count="0">Primary</button>

button("Secondary")

#> <button class="dull-button dull-input btn btn-secondary" type="button" role="button" data-count="0">Secondary</button>

button("Success", context = "success")

#> <button class="dull-button dull-input btn btn-success" type="button" role="button" data-count="0">Success</button>

button("Info", context = "info", outline = TRUE)

#> <button class="dull-button dull-input btn btn-outline-info" type="button" role="button" data-count="0">Info</button>

button("\u2715", context = "warning")

#> <button class="dull-button dull-input btn btn-warning" type="button" role="button" data-count="0">✕</button>

button("Danger!", context = "danger", disable = TRUE)

#> <button class="dull-button dull-input btn btn-danger" type="button" role="button" data-count="0">Danger!</button>

if (interactive()) {
  shinyApp(
    ui = container(
      listGroup(
        listGroupItem(
          inlineForm(
            button(
              id = "clicker1",
              "Simple button"
            )
          ),
          badge = badge(id = "badge1", 0)
        ),
        listGroupItem(
          inlineForm(
            button(
              id = "clicker2",
              label = "Click me!",
            )
          ),
          badge = badge(id = "badge2", 0)
        ),
        listGroupItem(
          button(
            id = "reset",
            label = "Reset",
            context = "primary"
          )
        )
      )
    ),
    server = function(input, output) {
      output$badge1 <- renderBadge(input$clicker1$count)

      output$badge2 <- renderBadge(
        value = {
          input$clicker2$count
        },
        context = {
          if (input$clicker2$count > 5) {
            "warning"
          } else {
            "default"
          }
        }
      )

      observeEvent(input$reset, {
        updateButton("clicker1", count = 0)
        updateButton("clicker2", count = 0)
      })
    }
  )

  shinyApp(
    ui = container(
      button(
        id = "group",
        label = c("First", "Second", "Third"),
        value = c("first", "second", "third")
      )
    ),
    server = function(input, output) {
      observe({
        print(input$group)
      })
    }
  )
}