Checkbox inputs and checkbox bars

A reactive input for selecting one or more values. A checkbox bar is a visually re-styling of a group of checkboxes. A checkbox bar appears as a set of buttons, but still acts as a checkbox group. Because checkbox bars are pseudo-buttons a visual context may be specified.

checkboxInput(labels = NULL, values = NULL, checked = NULL,
  inline = FALSE, bar = FALSE, context = "secondary", ..., id = NULL)

updateCheckbox(id, context, session = getDefaultReactiveDomain())

Arguments

labels
A character vector specifying the labels of the checkboxes, defaults to NULL.

values
A character vector specifying the values of the checkboxes, must be the same length as labels, defaults to NULL.

checked
A logical vector specifying which of the checkboxes render in a checked state, if specified must be the same length as labels, defaults to NULL. If unspecified all checkboxes render in an unchecked state.

inline
If TRUE, the checkbox options are rendered inline, otherwise the checkbox options appear on individual lines, defaults to FALSE.

bar
If TRUE, changes only the visual appearance of a group of checkboxes, checkboxes are rendered as a button bar or set of buttons, defaults to FALSE.

context
Applied only when bar = TRUE, one of "primary", "secondary", "success", "info", "warning", or "danger" specifying the visual context of the checkbox bar, defaults to "secondary".

...
Additional named arguments passed on as HTML attributes to the parent element.

id
A character string specifying the id of the checkbox bar, if specified a reactive input is available to the shiny server function, defaults to NULL.

Examples

stub

#> Error in eval(expr, envir, enclos): object 'stub' not found