Collapsible sections

collapse allows content to be hidden and toggled into a shown state. The collapsed content is toggled by a button. A collapsed section may be toggled between hidden and shown using toggleCollapse, shown with showCollapse, and hidden with hideCollapse.

collapse(..., label = NULL, context = "secondary")

showCollapse(id, session = getDefaultReactiveDomain())

hideCollapse(id, session = getDefaultReactiveDomain())

toggleCollapse(id, session = getDefaultReactiveDomain())

Arguments

...
Any number of tag elements to include in the collapsible section or additional named arguments passed as attributes to the parent element.

label
A character string specifying the label of the collapse, passed to button, defaults to NULL.

context
A character string specifying the visual context of the toggle button, passed to button, defaults to "secondary".

id
A character string specifying the id of a collapse section.

session
A session object passed to the shiny server function, defaults to shiny::getDefaultReactiveDomain().

Details

The showCollapse, hideCollapse, and toggleCollapse functions can be used very similarly to the update* functions found in shiny. Because collapse is not a reactive input a single updateCollapse function is not included with dull nor did it make sense to create a renderCollapse function as a collapse is not a reactive output.

Examples

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          collapse(
            id = "book",
            label = "The Time Machine",
            tags$div(
              class  = "card card-block",
              "\"The Time Traveller (for so it will be convenient to speak
              of him) was expounding a recondite matter to us. His grey eyes
              shone and twinkled, and his usually pale face was flushed and
              animated. The fire burned brightly, and the soft radiance of
              the incandescent lights in the lilies of silver caught the
              bubbles that flashed and passed in our glasses.\""
            )
          )
        ),
        col(
          button(id = "toggle", "Toggle collapse"),
          button(id = "hide", "Only hide", context = "warning"),
          button(id = "show", "Only show", context = "info")
        )
      )
    ),
    server = function(input, output) {
      observeEvent(input$toggle, {
        toggleCollapse("book")
      })

      observeEvent(input$hide, {
        hideCollapse("book")
      })

      observeEvent(input$show, {
        showCollapse("book")
      })
    }
  )
}