Color text and backgrounds

Utilities functions to change element text or background color. These utilities are intended to give visual context to blocks of text or elements. When using a dark background one may want to specify a lighter color for any element text.

utils$text(context)

utils$background(context)

Arguments

context
A character string specifying the visual context of a block of text or background color of an element.

For utils$text, one of "muted", "primary", "success", "info", "warning", "danger", or "white".

For utils$background, one of "muted", "primary", "success", "info", "warning", "danger", "inverse" or "faded".

See also

Other utilities: flexbox, float, hidden, position, sizing, spacing, viewport