Dropdown Menu

Dropdown menus function similar to a set of buttons. The initial dropdown open does not trigger a reactive event, but a subsequent click on one of the menu choices triggers a reactive event. The value of a dropdown item may be specified in dropdownItem, otherwise the default value is NULL.

dropdown(..., label = NULL, context = "secondary", split = FALSE,
  align = "left", dropup = FALSE)

dropdownItem(label = NULL, value = NULL, disabled = FALSE, ...)

dropdownDivider()

dropdownHeader(label)

Arguments

...
Any number of dropdown items passed to dropdown or additional named arguments passed as HTML attributes to the respective parent element.

label
A character string specifying the label for a dropdown, dropdown item, or dropdown header, defaults to NULL for dropdowns and dropdown items, but is required for a dropdown header.

context
One of "primary", "secondary", "success", "info", "warning", or "danger" specifying the visual context of the dropdown, defaults to "secondary".

split
If TRUE, the dropdown is changed aesthetically such that the dropdown icon is split into a separate, smaller button, defaults to FALSE.

align
One of "left" or "right" specifying which side of the dropdown button the menu aligns to, defaults to "left".

dropup
If TRUE, the dropdown menu extends upwards, defaults to FALSE.

value
A character string specifying the value of a dropdown item, defaults to NULL.

disabled
If TRUE, the dropdown item renders in a disabled state and will not trigger a reactive event, defaults to FALSE. The state of a dropdown item may be toggled with updateDropdownItem.

Examples

dropdown(
  label = "Secondary",
  dropdownItem("Option 1"),
  dropdownItem("Option 2")
)

#> <div class="dull-dropdown btn-group dropdown">
#>   <button class="dull-button dull-input btn btn-secondary dropdown-toggle" data-count="0" data-toggle="dropdown" role="button" type="button">Secondary</button>
#>   <div class="dropdown-menu">
#>     <a class="dull-dropdown-item dropdown-item">Option 1</a>
#>     <a class="dull-dropdown-item dropdown-item">Option 2</a>
#>   </div>
#> </div>

dropdown(
  id = "flavor",
  label = "Ice cream flavors",
  dropdownItem("Vanilla", "vanilla"),
  dropdownItem("Chocolate", "chocolate"),
  dropdownItem("Mint", "mint")
)

#> <div class="dull-dropdown btn-group dropdown" id="flavor">
#>   <button class="dull-button dull-input btn btn-secondary dropdown-toggle" data-count="0" data-toggle="dropdown" role="button" type="button">Ice cream flavors</button>
#>   <div class="dropdown-menu">
#>     <a class="dull-dropdown-item dropdown-item" data-value="vanilla">Vanilla</a>
#>     <a class="dull-dropdown-item dropdown-item" data-value="chocolate">Chocolate</a>
#>     <a class="dull-dropdown-item dropdown-item" data-value="mint">Mint</a>
#>   </div>
#> </div>

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          dropdown(
            id = "desserts",
            label = "Pick a dessert to learn more",
            align = "right",
            dropdownHeader("Frozen"),
            dropdownItem("Ice cream", "icecream"),
            dropdownItem("Gelato", "gelato"),
            dropdownHeader("Baked goods"),
            dropdownItem("Cake", "cake"),
            dropdownItem("Pie", "pie")
          )
        ),
        col(
          dropdown(
            id = "spices",
            label = "Check for a spice",
            dropdownItem("Cardamom", "cardamom"),
            dropdownItem("Nutmeg", "nutmeg"),
            dropdownItem("Vanilla", "vanilla")
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        print(input$desserts)
      })

      observe({
        print(input$spices)
      })
    }
  )
}