Flex box and flex item utilities

These utility functions are used to build flex layouts and are viewport responsive.

Convert tags into flex boxes and flex items. These functions may be chained together to specify different flex behavior for different viewports. Multi length arguments may be passed to flexbox and flexitem. The name of each argument should be one of small, sm, medium, md, large, lg, extra-large, lg or blank (extra-small) indicating a viewport size at which to apply the flex behavior.

utils$flexitem(align = NULL, margin = NULL, order = NULL, viewport = NULL)

utils$flexbox(direction = NULL, reverse = FALSE, justify = NULL,
  align = NULL, content = NULL, wrap = NULL, viewport = NULL)

flexbox(..., direction = NULL, reverse = FALSE, justify = NULL,
  align = NULL, content = NULL, wrap = NULL)

flexitem(align = NULL, margin = NULL, order = NULL, viewport = NULL)

Arguments

direction
One of "row" or "column" specifying the direction of a flex container's items, defaults to "row".

Row-wise flex containers handle elements from left to right in rows, the main axis is the x-axis, and the cross axis is the y-axis.

Column-wise flex containers handle elements from top to bottom in columns, the main axis is y-axis, and the cross axis is the x-axis.

reverse
If TRUE, the ordering of flex items inside a flex container is reversed and the items are aligned to the right side of the container, defaults to FALSE. This argument has no effect unless direction is specified.

justify
One "start", "end", "center", "between", or "around" specifying how a flex containers items are aligned along the main axis, defaults to NULL. Note that setting justify to "end" is not the same as setting reverse to TRUE. Unlike align, there is no flex item equivalent of justify.

align
One of "start", "end", "center", "baseline", or "stretch" specifying how a flex container aligns items or how a flex item aligns itself within a flex container along the cross axis, defaults to NULL. Note that, "start" and "baseline" are noticeably different only when the content or font size of flex items are not the same. This is because baseline aligns items according to the bottom of an item's content.

content
One of "start", "end", "center", "between", "around", or "stretch", defaults to NULL.

wrap
One of "nowrap", "wrap", or "reverse" specifying how flex items will wrap inside a flex container, defaults to NULL. If "nowrap", items are condensed so as to fit on a single row or column, possibly extending past the boundary of the parent flex container. If "wrap" or "reverse", items are not condensed and will wrap within the parent flex container onto new rows or columns.

margin
One of "left", "right", "top", or "bottom" specifying an auto margin for a flex item, defaults to NULL. On an aligned axis, a flex item with an auto margin will push away its sibling items. On a justified axis, a flex item with an auto margin will push itself away from siblings.

viewport
One of "small", "medium", "large", or "extra-large", specifying the minimum screen size at which specified styles are applied, defaults to NULL. Multiple util$* calls may be collated together to specify different layouts for element(s) depending on screen size. When viewport is NULL specified styles will apply to the smallest screens on up. Not all styles have this functionality.

viewport
One of "small", "medium", "large", or "extra-large", specifying the minimum screen size at which specified flex behavior is applied, defaults to NULL.

direction
One of "row" or "column" specifying the direction of a flex container's items, defaults to "row".

Row-wise flex containers handle elements from left to right in rows, the main axis is the x-axis, and the cross axis is the y-axis.

Column-wise flex containers handle elements from top to bottom in columns, the main axis is y-axis, and the cross axis is the x-axis.

reverse
If TRUE, the ordering of flex items inside a flex container is reversed and the items are aligned to the right side of the container, defaults to FALSE. This argument has no effect unless direction is specified.

justify
One "start", "end", "center", "between", or "around" specifying how a flex containers items are aligned along the main axis, defaults to NULL. Note that setting justify to "end" is not the same as setting reverse to TRUE. Unlike align, there is no flex item equivalent of justify.

align
One of "start", "end", "center", "baseline", or "stretch" specifying how a flex container aligns items or how a flex item aligns itself within a flex container along the cross axis, defaults to NULL. Note that, "start" and "baseline" are noticeably different only when the content or font size of flex items are not the same. This is because baseline aligns items according to the bottom of an item's content.

content
One of "start", "end", "center", "between", "around", or "stretch", defaults to NULL.

wrap
One of "nowrap", "wrap", or "reverse" specifying how flex items will wrap inside a flex container, defaults to NULL. If "nowrap", items are condensed so as to fit on a single row or column, possibly extending past the boundary of the parent flex container. If "wrap" or "reverse", items are not condensed and will wrap within the parent flex container onto new rows or columns.

margin
One of "left", "right", "top", or "bottom" specifying an auto margin for a flex item, defaults to NULL. On an aligned axis, a flex item with an auto margin will push away its sibling items. On a justified axis, a flex item with an auto margin will push itself away from siblings.

How it all works

reverse

Flex container with reverse = FALSE.

| Item 1 | Item 2 | Item 3 | ============= |

Flex container with reverse = TRUE

| ============= | Item 3 | Item 2 | Item 1 |

justify

Flex container with justify = "start"

| Item 1 | Item 2 | Item 3 | ============= |

Flex container with justify = "end"

| ============= | Item 1 | Item 2 | Item 3 |

Flex container with justify = "center"

| ===== | Item 1 | Item 2 | Item 3 | ===== |

Flex container with justify = "between"

| Item 1 | ===== | Item 2 | ===== | Item 3 |

Flex container with justify = "around"

| = | Item 1 | = | Item 2 | = | Item 3 | = |

align (flexbox)

Parent flex container with align = "start"

| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |
|        |        |        |               |

Parent flex container with align = "end"

|        |        |        |               |
|        |        |        |               |
| Item 1 | Item 2 | Item 3 | ============= |

Parent flex container with align = "center"

|        |        |        |               |
| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |

Parent flex container with align = "baseline"

| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |
|        |        |        |               |

Parent flex container with align = "stretch"

| It     | It     | It     | ============= |
|   em   |   em   |   em   |               |
|      1 |      2 |      3 |               |

align (flexitem)

Parent flex container with align = "stretch", child Item 2 with align = "start"

| It     | Item 2 | It     | ============= |
|   em   |        |   em   |               |
|      1 |        |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "end"

| It     |        | It     | ============= |
|   em   |        |   em   |               |
|      1 | Item 2 |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "center"

| It     |        | It     | ============= |
|   em   |        |   em   |               |
|      1 | Item 2 |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "baseline"

| It     | Item 2 | It     | ============= |
|   em   |        |   em   |               |
|      1 |        |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "stretch"

| It     | It     | It     | ============= |
|   em   |   em   |   em   |               |
|      1 |      2 |      3 |               |

margin

Parent flex container with justify = "end", child Item 1 with margin = "right" pushes self away from siblings

| Item 1 | ============= | Item 2 | Item 3 |

Parent flex container with align = "start", child Item 1 with margin = "right" pushes siblings away from self

| Item 1 | ============= | Item 2 | Item 3 |

Parent flex container with justify = "start", child Item 3 with margin = "left" pushes self away from siblings

| Item 1 | Item 2 | ============= | Item 3 |

Parent flex container with align = "end", child Item 1 with margin = "left" pushes siblings away from self

| Item 1 | Item 2 | ============= | Item 3 |

wrap

Parent flex container with wrap = "nowrap"

| Item | Item | Item | Item | Item | Item |
| 1    | 2    | 3    | 4    | 5    | 6    |

Parent flex container with wrap = "wrap"

| Item 1 | Item 2 | Item 3 | Item 4 | === |
| Item 5 | Item 6 |                       |

Parent flex container with wrap = "reverse"

| Item 5 | Item 6 |                       |
| Item 1 | Item 2 | Item 3 | Item 4 | === |

Better understanding flex behavior

reverse

Flex container with reverse = FALSE.

| Item 1 | Item 2 | Item 3 | ============= |

Flex container with reverse = TRUE

| ============= | Item 3 | Item 2 | Item 1 |

justify

Flex container with justify = "start"

| Item 1 | Item 2 | Item 3 | ============= |

Flex container with justify = "end"

| ============= | Item 1 | Item 2 | Item 3 |

Flex container with justify = "center"

| ===== | Item 1 | Item 2 | Item 3 | ===== |

Flex container with justify = "between"

| Item 1 | ===== | Item 2 | ===== | Item 3 |

Flex container with justify = "around"

| = | Item 1 | = | Item 2 | = | Item 3 | = |

align (flexbox)

Parent flex container with align = "start"

| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |
|        |        |        |               |

Parent flex container with align = "end"

|        |        |        |               |
|        |        |        |               |
| Item 1 | Item 2 | Item 3 | ============= |

Parent flex container with align = "center"

|        |        |        |               |
| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |

Parent flex container with align = "baseline"

| Item 1 | Item 2 | Item 3 | ============= |
|        |        |        |               |
|        |        |        |               |

Parent flex container with align = "stretch"

| It     | It     | It     | ============= |
|   em   |   em   |   em   |               |
|      1 |      2 |      3 |               |

align (flexitem)

Parent flex container with align = "stretch", child Item 2 with align = "start"

| It     | Item 2 | It     | ============= |
|   em   |        |   em   |               |
|      1 |        |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "end"

| It     |        | It     | ============= |
|   em   |        |   em   |               |
|      1 | Item 2 |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "center"

| It     |        | It     | ============= |
|   em   |        |   em   |               |
|      1 | Item 2 |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "baseline"

| It     | Item 2 | It     | ============= |
|   em   |        |   em   |               |
|      1 |        |      3 |               |

Parent flex container with align = "stretch", child Item 2 with align = "stretch"

| It     | It     | It     | ============= |
|   em   |   em   |   em   |               |
|      1 |      2 |      3 |               |

margin

Parent flex container with justify = "end", child Item 1 with margin = "right" pushes self away from siblings

| Item 1 | ============= | Item 2 | Item 3 |

Parent flex container with align = "start", child Item 1 with margin = "right" pushes siblings away from self

| Item 1 | ============= | Item 2 | Item 3 |

Parent flex container with justify = "start", child Item 3 with margin = "left" pushes self away from siblings

| Item 1 | Item 2 | ============= | Item 3 |

Parent flex container with align = "end", child Item 1 with margin = "left" pushes siblings away from self

| Item 1 | Item 2 | ============= | Item 3 |

wrap

Parent flex container with wrap = "nowrap"

| Item | Item | Item | Item | Item | Item |
| 1    | 2    | 3    | 4    | 5    | 6    |

Parent flex container with wrap = "wrap"

| Item 1 | Item 2 | Item 3 | Item 4 | === |
| Item 5 | Item 6 |                       |

Parent flex container with wrap = "reverse"

| Item 5 | Item 6 |                       |
| Item 1 | Item 2 | Item 3 | Item 4 | === |

See also

For more information about bootstrap HTML utility classes please refer to the online reference page.

For more information about bootstrap HTML utility classes please refer to the online reference page.

Other utilities: colors, float, hidden, position, sizing, spacing, viewport

Examples

# stub

if (interactive()) {
  shinyApp(
    ui = container(

    ),
    server = function(input, output) {

    }
  )
}