Group inputs, combination buttons and text

An input group is composite reactive input which may consist of one or two buttons, dropdowns, character addons, or any combination of these elements. Character addons, specified with left and right may be used to ensure an input group's value always has a certain prefix or suffix and render a visual cue to indicate this behavior. Buttons and dropdowns may be included to control when the input group's reactive value updates. See below for more information and examples.

groupInput(placeholder = NULL, value = NULL, left = NULL, right = NULL,
  ..., id = NULL)

Arguments

placeholder
A character string specifying placeholder text for the input group, defaults to NULL.

value
A character string specifying an initial value for the input group, defaults to NULL.

left, right
A character string, button element, or dropdown element, used as the left or right addon, respectively, of the input group, both default to NULL. Addon's affect the reactive events and value of the input group, see the details section below for more information.

id
A character string specifying an id for the input, defaults to NULL. When specified, a reactive value, input$<id>, is available to the shiny server function.

Details

reactive event

If either left or right is a button or a dropdown the reactive value of the input group is changed when either button or a dropdown item is clicked. If either left and right are character strings or NULL the reactive value of the input group will update each time the text input is changed by the user. input value

An input group's value is NULL when its text input is empty and neither, if any, buttons or dropdown item's have been clicked. Otherwise, a input group's value is list of two named elements. The first element is value which is the concatenation of left, if a character string, the text input value, and right, if a character string. The second element is click and is indicates which button or dropdown item was clicked. The value of click depends on the value of the button or dropdown item, see button or dropdownItem.

See also

Other inputs: button, radioInput, selectInput, textInput

Examples

if (interactive()) {
  shinyApp(
    ui = container(
      groupInput(
        id = "nearlytext"
      ),
      groupInput(
        id = "username",
        placeholder = "username",
        left = "@",
        right = button(
          label = "Go!",
          value = "go"
        )
      ),
      groupInput(
        id = "github",
        left = "https://github.com/"
      ),
      groupInput(
        id = "preference",
        left = button(
          label = fontAwesome("thumbs-up"),
          value = "up"
        ),
        right = button(
          label = fontAwesome("thumbs-down"),
          value = "down"
        )
      ),
      groupInput(
        id = "options",
        placeholder = "",
        left = dropdown(
          label = "Action",
          dropdownItem("Action", "action"),
          dropdownItem("Another action", "another"),
          dropdownItem("Something else", "something"),
          dropdownDivider(),
          dropdownItem("Final", "final")
        )
      )
    ),
    server = function(input, output) {
      observe({
        print(input$nearlytext)
      })

      observe({
        print(input$username)
      })

      observe({
        print(input$github)
      })

      observe({
        print(input$preference)
      })

      observe({
        print(input$options)
      })
    }
  )
}