Responsive Images and Figures

A small update to <img> and <figure>. Additional arguments has been added to the img builder function. Figures have specific arguments for an img child element and caption.

img(..., fluid = TRUE, align = NULL, thumbnail = FALSE)

figure(..., image = NULL, caption = NULL, align = "left")

Arguments

fluid
If TRUE, the image will scale with its parent element, defaults to TRUE.

align
For img, one of "left" or "right" specifying which side of the parent element the image should align to, defaults to NULL in which case no alignment is applied.

For figure, one of "left", "right", or "center" specifying which side of the figure the figure caption is aligned to, defaults to "left".

thumbnail
If TRUE, the image is rendered with a small border, defaults to FALSE.

image
An img element, it is best to leave the image as fluid, defaults to NULL.

caption
A character vector specifying a caption for the image, defaults to NULL.

See also

Other content: blockquote

Examples

if (interactive()) {
  shinyApp(
    ui = container(
      # A figure with a,
      #   - fluid image
      #   - left-aligned caption
      figure(
        image = img(src = "http://bit.ly/2qchbEB"),
        caption = "Stock cat photo."
      )
    ),
    server = function(input, output) {

    }
  )
}