List Group

A way of handling and outlining content as a list.

listGroup(...)

listGroupItem(label = NULL, value = NULL, context = NULL,
  active = FALSE, disabled = FALSE, hover = FALSE, badge = NULL, ...)

updateListGroupItem(id, label = NULL, value = NULL, context = NULL,
  active = NULL, disabled = NULL, session = getDefaultReactiveDomain())

Arguments

...
List items, built with listGroupItem, character strings, custom HTML content, or named arguments passed as HTML attributes to the parent div.

label
List item label, defaults to NULL.

value
The value of the list item, defaults to NULL.

context
A character string specifying the visual context of the list item, one of "success", "info", "warning", or "danger", defaults to NULL.

active
If TRUE the list item is rendered in the active state, defaults to FALSE.

disabled
If TRUE the list item is rendered in the disabled state, disabled items do not receive click events, defaults to FALSE.

hover
If TRUE the list item receives additional hover animation, under the hood the list item is a <a> element instead of <li>, defaults to FALSE.

badge
A badge which is right aligned inside the list item, see renderBadge for more on how to work with badges, defaults to NULL.

Details

Character strings are converted into list items. Custom tags are left as is and must include Bootstrap classes ahead of time. To specify a value for a list item please include a data-value attribute in the .list-group-item element.

See also

For more information on Bootstrap list groups please refer to the reference page.

Examples

listGroup(
  "Item 1",
  "Item 2",
  "Item 3"
)

#> <ul class="dull-list-group list-group list-group-flush">
#>   <li class="dull-list-group-item list-group-item" data-value="Item 1">Item 1</li>
#>   <li class="dull-list-group-item list-group-item" data-value="Item 2">Item 2</li>
#>   <li class="dull-list-group-item list-group-item" data-value="Item 3">Item 3</li>
#> </ul>

# highlight list item
listGroup(
  "Item 1",
  listGroupItem("Item 2", active = TRUE),
  "Item 3"
)

#> <ul class="dull-list-group list-group list-group-flush">
#>   <li class="dull-list-group-item list-group-item" data-value="Item 1">Item 1</li>
#>   <li class="dull-list-group-item list-group-item active">Item 2</li>
#>   <li class="dull-list-group-item list-group-item" data-value="Item 3">Item 3</li>
#> </ul>

# disable list items
listGroup(
  listGroupItem("Item 1", disabled = TRUE),
  "Item 2",
  listGroupItem("Item 3", disabled = TRUE)
)

#> <ul class="dull-list-group list-group list-group-flush">
#>   <li class="dull-list-group-item list-group-item disabled">Item 1</li>
#>   <li class="dull-list-group-item list-group-item" data-value="Item 2">Item 2</li>
#>   <li class="dull-list-group-item list-group-item disabled">Item 3</li>
#> </ul>

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          width = 4,
          tags$h4("Any preferences?"),
          listGroup(
            id = "preferences",
            listGroupItem("No heights", "noheights"),
            listGroupItem("No animals", "noanimals")
          )
        ),
        col(
          tags$h4("Possible outings"),
          listGroup(
            listGroupItem(
              id = "optZoo",
              "Zoo visit!"
            ),
            listGroupItem(
              id = "optSkydive",
              "Sky diving!"
            ),
            listGroupItem(
              id = "optPern",
              "Visit Pern"
            )
          )
        )
      )
    ),
    server = function(input, output) {
      observe({
        if ("noheights" %in% input$preferences) {
          updateListGroupItem("optSkydive", context = "warning")
        }
      })

    }
  )
}