Page Navigation

Create a nav.

nav(..., align = "left", width = NULL, vertical = FALSE, tabs = FALSE,
  pills = FALSE)

navItem(label = NULL, active = FALSE, disabled = FALSE, ...)

navDropdown(label = NULL, ...)

Arguments

...
navItems or navDropdowns passed to nav, dropdownItems passed to navDropdown, or named arguments passed as HTML attributes to the respective parent element.

align
One of "left", "right", "center", or "horizontal" specifying the alignment of the nav, defaults to "left".

width
If specified, one of "fill" or "justify" specifying how the nav will fill its parent element, defaults to NULL. If "justify" all nav links have equal width, when "fill" nav links are only equally spaced.

vertical
If TRUE, the nav is rendered vertically, defaults to FALSE.

tabs
If TRUE, the nav is rendered as a set of tabs for a tabbed interface, defaults to FALSE.

pills
If TRUE, navItems are rendered as pills, defaults to FALSE.

label
A character vector specifying the label of a nav item, defaults to NULL.

active
If TRUE, the nav item renders in a an active state, defaults to FALSE.

disabled
If TRUE, the nav item renders in a disabled state and will not receive click events, defaults to FALSE.

Details

To center a horizontal nav specify align = "horiztonal" and width = "fill" or width = "justify".

See also

For more information about the Bootstrap nav please follow this link.

Examples

nav(
  navItem("Home", href = "#home"),
  navItem("About", href = "#about")
)

#> <ul class="nav" role="navigation">
#>   <li class="nav-item">
#>     <a class="nav-link">Home</a>
#>   </li>
#>   <li class="nav-item">
#>     <a class="nav-link">About</a>
#>   </li>
#> </ul>

# active link
nav(
  navItem("Home", href = "#home", active = TRUE),
  navItem("About", href = "#about")
)

#> <ul class="nav" role="navigation">
#>   <li class="nav-item">
#>     <a class="nav-link active">Home</a>
#>   </li>
#>   <li class="nav-item">
#>     <a class="nav-link">About</a>
#>   </li>
#> </ul>

# active tab
nav(
  tabs = TRUE,
  navItem("Home", href = "#home", active = TRUE),
  navItem("About", href = "#about"),
  navItem("Blog", href = "#blog")
)

#> <ul class="nav nav-tabs" role="navigation tablist">
#>   <li class="nav-item">
#>     <a class="nav-link active">Home</a>
#>   </li>
#>   <li class="nav-item">
#>     <a class="nav-link">About</a>
#>   </li>
#>   <li class="nav-item">
#>     <a class="nav-link">Blog</a>
#>   </li>
#> </ul>

if (interactive()) {
  shinyApp(
    ui = container(
      nav(
        navItem(
          label = "Home",
          active = TRUE
        ),
        navItem(
          label = "Posts"
        ),
        navItem(
          label = "About"
        ),
        navDropdown(
          dropdownItem()
        )
      )
    ),
    server = function(input, output) {

    }
  )
}