Affix elements

These utility functions are used to affix elements to the top or bottom of a page. When the page is scrolled affixed elements will remain on the top or bottom. A sticky element is affixed to the top of the page once the page scrolls past it. Note that the CSS used in the sticky functionality is not fully supported by all browsers.

utils$position(fixed)

Arguments

fixed
One of "top", "bottom", or "sticky" positioning and fixing an element to the top or bottom of the viewport. If "sticky", the element is only fixed to the top of the viewport after scrolling past it.

See also

Other utilities: colors, flexbox, float, hidden, sizing, spacing, viewport