Element Spacing

Use these utilities to add responsive padding or margins to a tag element. Both margin and padding are viewport responsive. Margin and padding sizes are specified relative to the font size of a page's root <html> element.

utils$margin(size, side = "all", viewport = NULL)

utils$padding(size, side = "all", viewport = NULL)

Arguments

size
An integer, 0 through 5, specifying the relative adjustment of an element's margin or padding, where 3 is the default amount of spacing, 1 is 0.25 times the default, and 5 is 3 times the default. 0 removes margins or padding.

side
One of "top", "left", "bottom", "right", or "all" specifying which side(s) to apply spacing to, defaults to "all".

viewport
One of "small", "medium", "large", or "extra-large", specifying the screen size at which the style is applied, defaults to NULL.

See also

The complete list of utilities. For more on viewport responsiveness see viewport.

Other utilities: colors, flexbox, float, hidden, position, sizing, viewport