Stream notifications

These functionalities allow an application to convey text progress updates to the user during long running processes.

streamOutput(id, ...)

updateStream(id, message, template = listGroupItem(),
  session = getDefaultReactiveDomain())

Arguments

id
A character string specifying the id of the stream output.

message
A character string specifying an update message.

template
A call listGroupItem which is used as a template for the message sent, any body element(s) or text of the item are replaced by message, defaults to listGroupItem().

session
The shiny session, defaults to the default reactive domain.

Examples

if (interactive()) {
  shinyApp(
    ui = container(
      button(id = "trigger", "Go!"),
      tags$div(
        id = "myStream"
      )
    ),
    server = function(input, output, session) {
      observeEvent(input$trigger, {
        for (i in seq_len(5)) {
          updateStream(
            id = "myStream",
            message = paste("Update:", i),
            template = listGroupItem(
              context = if (i %% 2) "warning" else "info"
            )
          )
          Sys.sleep(1);
        }
      })
    }
  )
}