Table output

Render a table.

tableThruput(id, borders = FALSE, compact = FALSE, invert = FALSE,
  striped = FALSE, hoverable = FALSE, context = NULL, responsive = TRUE,
  ...)

renderTable(expr, numbered = FALSE, env = parent.frame(), quoted = FALSE)

Arguments

id
A character string specifying the id of the table output.

borders
If TRUE, the table renders with borders, defaults to FALSE.

compact
If TRUE, table cell padding is cut in half to reduce the size of the table, defaults to FALSE.

invert
If TRUE, the table renders with a dark background, light text, and any borders are light, defaults to FALSE.

striped
If TRUE, the table renders with a striped pattern, defaults to FALSE.

hoverable
If TRUE, each table row has a hover state, defaults to FALSE.

context
One "success", "info", "warning", "danger", specifying the context of selected table rows, defaults to NULL, in which case selected rows are highlighted in grey.

responsive
If TRUE, the table will scroll horizontally on small viewports, viewports under 768 pixels, defaults to TRUE. There is no effect when the viewport height is greater than 768 pixels.

...
Additional named arguments passed as HTML attributes to the parent element.

Examples

if (interactive()) {
  shinyApp(
    ui = container(
      row(
        col(
          tableThruput(
            id = "tbl",
            hoverable = TRUE,
            context = "warning",
            borders = TRUE
          )
        ),
        col(
          tableThruput(
            id = "subset",
            borders = TRUE
          )
        )
      )
    ),
    server = function(input, output) {
      output$tbl <- renderTable(
        numbered = TRUE,
        iris[1:10, ]
      )

      output$subset <- renderTable(
        numbered = TRUE,
        input$tbl
      )
    }
  )
}