Add Tooltips and Popovers

Both functions are a means of adding help text to HTML components. Tooltips appear on hover and popovers appear when the element is clicked.

tooltip(content, text, placement = "top")

popover(content, text, placement = "top")

Arguments

content
The tag to add a tooltip or popover to.

text
The tooltip or popover text.

placement
A character vector specifying where the tooltip or popover is placed, defaults to top, one of "left", "top", "right", or "bottom".

Examples

tooltip(
  button("My button!"),
  text = "This is my button",
)

#> <button class="dull-button dull-input btn btn-secondary" type="button" role="button" data-count="0" data-toggle="tooltip" data-placement="top" title="This is my button">
#>   My button!
#>   head
#>   <script src="dull/tether/tether.min.js"></script>
#>   <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>   <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>   <script src="dull/js/dull.min.js"></script>
#>   <link rel="stylesheet" href="dull/css/dull.css"/>
#> </button>

tooltip(
  button("Another button"),
  text = "I won't do much when clicked, sorry",
  placement = "bottom"
)

#> <button class="dull-button dull-input btn btn-secondary" type="button" role="button" data-count="0" data-toggle="tooltip" data-placement="bottom" title="I won&#39;t do much when clicked, sorry">
#>   Another button
#>   head
#>   <script src="dull/tether/tether.min.js"></script>
#>   <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>   <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>   <script src="dull/js/dull.min.js"></script>
#>   <link rel="stylesheet" href="dull/css/dull.css"/>
#> </button>

listGroup(
  tooltip(
    listGroupItem("Item 1"),
    "This is the first item"
  ),
  tooltip(
    listGroupItem("Item 2"),
    "This is the second item"
  )
)

#> <ul class="dull-list-group list-group list-group-flush">
#>   <li class="dull-list-group-item list-group-item" data-toggle="tooltip" data-placement="top" title="This is the first item">
#>     Item 1
#>     head
#>     <script src="dull/tether/tether.min.js"></script>
#>     <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>     <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>     <script src="dull/js/dull.min.js"></script>
#>     <link rel="stylesheet" href="dull/css/dull.css"/>
#>   </li>
#>   <li class="dull-list-group-item list-group-item" data-toggle="tooltip" data-placement="top" title="This is the second item">
#>     Item 2
#>     head
#>     <script src="dull/tether/tether.min.js"></script>
#>     <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>     <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>     <script src="dull/js/dull.min.js"></script>
#>     <link rel="stylesheet" href="dull/css/dull.css"/>
#>   </li>
#> </ul>

popover(
  button("Click me!"),
  "This text appears on click"
)

#> <button class="dull-button dull-input btn btn-secondary" type="button" role="button" data-count="0" data-container="body" data-toggle="popover" data-placement="top" data-content="This text appears on click">
#>   Click me!
#>   head
#>   <script src="dull/tether/tether.min.js"></script>
#>   <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>   <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>   <script src="dull/js/dull.min.js"></script>
#>   <link rel="stylesheet" href="dull/css/dull.css"/>
#> </button>

# list items with popovers
listGroup(
  popover(
    listGroupItem("Some text"),
    "The first list item"
  ),
  popover(
    listGroupItem("More text"),
    "The second list item"
  ),
  popover(
    listGroupItem("Final text"),
    "Did you find this text?"
  )
)

#> <ul class="dull-list-group list-group list-group-flush">
#>   <li class="dull-list-group-item list-group-item" data-container="body" data-toggle="popover" data-placement="top" data-content="The first list item">
#>     Some text
#>     head
#>     <script src="dull/tether/tether.min.js"></script>
#>     <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>     <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>     <script src="dull/js/dull.min.js"></script>
#>     <link rel="stylesheet" href="dull/css/dull.css"/>
#>   </li>
#>   <li class="dull-list-group-item list-group-item" data-container="body" data-toggle="popover" data-placement="top" data-content="The second list item">
#>     More text
#>     head
#>     <script src="dull/tether/tether.min.js"></script>
#>     <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>     <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>     <script src="dull/js/dull.min.js"></script>
#>     <link rel="stylesheet" href="dull/css/dull.css"/>
#>   </li>
#>   <li class="dull-list-group-item list-group-item" data-container="body" data-toggle="popover" data-placement="top" data-content="Did you find this text?">
#>     Final text
#>     head
#>     <script src="dull/tether/tether.min.js"></script>
#>     <link rel="stylesheet" href="dull/bootstrap/css/bootstrap.min.css"/>
#>     <script src="dull/bootstrap/js/bootstrap.min.js"></script>
#>     <script src="dull/js/dull.min.js"></script>
#>     <link rel="stylesheet" href="dull/css/dull.css"/>
#>   </li>
#> </ul>